refactor: move types into types.ts
