+++ package.json
@@ -28,8 +28,8 @@
     "clean": "rm -rf ./lib",
     "lint": "npx eslint .",
     "prettify": "npx prettier --write .",
-    "build:types": "tsc --declaration --outDir ./lib --emitDeclarationOnly",
-    "build:cjs": "tsc -p tsconfig.cjs.json && mv ./lib/cjs/index.js ./lib/cjs/index.cjs && cp ./lib/index.d.ts ./lib/index.d.cts",
+    "build:types": "npx tsup src/index.ts --out-dir lib --dts-only --format esm,cjs",
+    "build:cjs": "tsc -p tsconfig.cjs.json && mv ./lib/cjs/index.js ./lib/cjs/index.cjs",
     "build:esm": "tsc -p tsconfig.json",
     "build": "npm run build:types && npm run build:esm && npm run build:cjs",
     "build:clean": "npm run clean && npm run build",
@@ -67,6 +67,7 @@
     "prettier": "^3.1.1",
     "standard-version": "^9.5.0",
     "ts-node": "^10.9.2",
+    "tsup": "^8.0.1",
     "typescript": "^5.3.3"
   },
   "dependencies": {
